requirements.txt
aiofile==3.8.1
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
caio==0.9.8
camelot-py==0.9.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
datefinder==0.7.3
decorator==5.1.1
deepdiff==5.8.1
deepl==1.11.0
distro==1.8.0
et-xmlfile==1.1.0
executing==1.1.0
frozenlist==1.3.1
idna==3.4
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
lxml==4.9.1
matplotlib-inline==0.1.6
mnemonic==0.20
multidict==6.0.2
numpy==1.23.3
opencv-python==4.6.0.66
openpyxl==3.0.10
ordered-set==4.1.0
pandas==1.5.0
parso==0.8.3
pdfminer.six==20220524
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyPDF2==2.11.1
python-dateutil==2.8.2
pytz==2022.4
regex==2022.9.13
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.1
tabula-py==2.5.1
toml==0.10.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
xmltodict==0.13.0
yarl==1.8.1
